Finished the Single Precison one for vectorization. with data distribution 1: 2: 2.
	modified:   SP/add.c
	modified:   SP/error.c
	modified:   SP/exact_rhs.c
	modified:   SP/exact_solution.c
	modified:   SP/header.h
	modified:   SP/initialize.c
	modified:   SP/ninvr.c
	modified:   SP/pinvr.c
	modified:   SP/rhs.c
	deleted:    SP/rhs.c.113t.vect
	modified:   SP/set_constants.c
	modified:   SP/sp.c
	modified:   SP/txinvr.c
	modified:   SP/tzetar.c
	modified:   SP/verify.c
	modified:   SP/x_solve.c
	modified:   SP/y_solve.c
	modified:   SP/z_solve.c
	modified:   config/make.def

+++ SP/exact_solution.c
@@ -36,7 +36,7 @@
 //---------------------------------------------------------------------
 // this function returns the exact solution at point xi, eta, zeta  
 //---------------------------------------------------------------------
-void exact_solution(double xi, double eta, double zeta, double dtemp[5])
+void exact_solution(float xi, float eta, float zeta, float dtemp[5])
 {
   int m;
